+++ app/Main.hs
@@ -14,27 +14,48 @@ module Main
     )
   where
 
-import Control.Applicative ((<**>), (<*>), (<|>), many, optional, some)
+import Prelude (fromIntegral)
+
+import Control.Applicative ((<**>), (<*>), (<|>), many, optional, pure, some)
 import Control.Exception (throwIO)
-import Control.Monad ((>>=))
+import Control.Monad ((>>=), guard)
+import Control.Monad.Fail (fail)
+import Data.Bool (Bool(False), not, otherwise)
+import Data.Eq ((==))
+import Data.Foldable (any, for_, null, traverse_)
 import Data.Function (($), (.))
-import Data.Functor ((<$>))
-import Data.Maybe (Maybe(Just, Nothing))
+import Data.Functor ((<$), (<$>), (<&>))
+import qualified Data.List as List (concat, elem, filter, sort, take)
+import qualified Data.List.NonEmpty as NonEmpty (toList)
+import Data.Maybe (Maybe(Just, Nothing), maybe)
 import Data.Semigroup ((<>))
 import Data.String (String, fromString)
 import Data.Version (makeVersion)
+import Data.Word (Word)
 import System.Environment (lookupEnv)
 import System.Exit (exitSuccess)
 import System.IO (FilePath, IO)
 
+import Data.CaseInsensitive (CI)
 import qualified Data.Either.Validation as Validation
-  ( Validation(Failure, Success)
-  )
+    ( Validation(Failure, Success)
+    )
 import Data.Text (Text)
+import qualified Data.Text as Text
+    ( drop
+    , intercalate
+    , isPrefixOf
+    , length
+    , uncons
+    , unlines
+    )
+import qualified Data.Text.IO as Text (putStr, putStrLn)
 import qualified Dhall (Decoder(Decoder, expected), input, inputFile)
 import qualified Options.Applicative as Options
     ( InfoMod
     , Parser
+    , ReadM
+    , auto
     , eitherReader
     , execParser
     , flag'
@@ -44,22 +65,32 @@ import qualified Options.Applicative as Options
     , help
     , helper
     , info
+    , internal
     , long
     , metavar
     , option
     , short
+    , str
     , strArgument
     , strOption
     )
 import qualified Prettyprinter (pretty, line)
 import qualified Prettyprinter.Render.Terminal as Prettyprinter (putDoc)
+import Safe (atDef, initMay, lastDef, lastMay)
+import qualified Database.SQLite.Simple as SQLite (withConnection)
 import System.Directory
     ( XdgDirectory(XdgConfig)
     , doesFileExist
     , getXdgDirectory
     )
 
-import Configuration (decodeConfiguration, mkDefConfiguration)
+import Configuration
+    ( Configuration(Configuration, sources)
+    , Source(Source, name)
+    , decodeConfiguration
+    , getCacheDirectory
+    , mkDefConfiguration
+    )
 import Client
     ( Action(ClearCache, List, Render, Update)
     , SomePlatform
@@ -68,6 +99,7 @@ import Client
     )
 import Version (VersionInfo(..), prettyVersionInfo)
 import Locale (Locale, parseLocale)
+import qualified Index (getCommands, getIndexFile, getLocales, getPlatforms)
 
 import Paths_tldr_client (version)
 
@@ -75,6 +107,7 @@ import Paths_tldr_client (version)
 main :: IO ()
 main = do
     (config, action) <- parseOptions decodeConfiguration mkDefConfiguration
+        completer
     client config action
 
 data Mode
@@ -92,6 +125,11 @@ data Mode
     | Version
     -- ^ Print version information and exit instead of doing anything.
     -- Configuration will not be parsed or used in any way.
+    | CompletionInfo
+    -- ^ Describe how command-line completion works in the form of a Dhall
+    -- expression.
+    | Completion (Maybe Text) Shell (Maybe Word) [Text]
+    -- ^ Do command-line completion.
 
 -- | Parse command line options, handle everything that is not related to the
 -- main purpose of this binary or return `Configuration` otherwise.
@@ -123,8 +161,10 @@ parseOptions
     -> IO config
     -- ^ Construct default configuration if there is no configuration
     -- available.
+    -> (config -> Shell -> Maybe Word -> [Text] -> IO ())
+    -- ^ Command-line completer.
     -> IO (config, Action)
-parseOptions decoder@Dhall.Decoder{expected} mkDef = do
+parseOptions decoder@Dhall.Decoder{expected} mkDef completer' = do
     configFile <- getXdgDirectory XdgConfig "tldr/config.dhall"
 
     parseOptions' configFile >>= \case
@@ -146,6 +186,35 @@ parseOptions decoder@Dhall.Decoder{expected} mkDef = do
                 { clientVersion = version
                 , tldrClientSpecificationVersion = makeVersion [1,5]
                 }
+            exitSuccess
+
+        CompletionInfo -> do
+            Text.putStrLn $ Text.unlines
+                [ "let toWordOptions ="
+                , "      λ(words : List Text) →"
+                , "        List/fold"
+                , "          Text"
+                , "          words"
+                , "          (List Text)"
+                , "          (λ(w : Text) → λ(ws : List Text) →\
+                                \ [ \"--word=${w}\" ] # ws)"
+                , "          ([] : List Text)"
+                , ""
+                , "in  λ(shell : < Bash | Fish | Zsh >) →"
+                , "    λ(index : Natural) →"
+                , "    λ(words : List Text) →"
+                , "        [ \"--completion\""
+                , "        , \"--index=${Natural/show index}\""
+                , "        , \"--shell=${merge { Bash = \"bash\",\
+                                \ Fish = \"fish\", Zsh = \"zsh\" } shell}\""
+                , "        ]"
+                , "      # toWordOptions words"
+                ]
+            exitSuccess
+
+        Completion possiblyConfig shell index words -> do
+            config <- parseConfig configFile possiblyConfig
+            completer' config shell index words
             exitSuccess
   where
     parseConfig :: FilePath -> Maybe Text -> IO config
@@ -195,7 +264,13 @@ parseOptions decoder@Dhall.Decoder{expected} mkDef = do
     options =
         versionFlag
         <|> printTypeFlag
+        <|> completionInfoFlag
         <|> (   ( typecheckFlag
+                <|> ( completionFlag
+                    <*> shellOption
+                    <*> optional indexOption
+                    <*> many wordOption
+                    )
                 <|> (   ( updateFlag
                         <|> (   ( listFlag
                                 <|> clearCacheFlag
@@ -220,30 +295,6 @@ parseOptions decoder@Dhall.Decoder{expected} mkDef = do
         -> Mode
     renderAction cmds lang platform sources cfg =
         Execute cfg (Render platform lang sources cmds)
-
-    listAction
-        :: Maybe Locale
-        -> Maybe SomePlatform
-        -> [Text]
-        -> Maybe Text
-        -> Mode
-    listAction lang platform sources cfg =
-        Execute cfg (List platform lang sources)
-
-    clearCacheAction
-        :: Maybe Locale
-        -> Maybe SomePlatform
-        -> [Text]
-        -> Maybe Text
-        -> Mode
-    clearCacheAction lang platform sources cfg =
-        Execute cfg (ClearCache platform lang sources)
-
-    updateAction
-        :: [Text]
-        -> Maybe Text
-        -> Mode
-    updateAction sources cfg = Execute cfg (Update sources)
 
     configOption :: Options.Parser Text
     configOption = Options.strOption
@@ -290,6 +341,9 @@ parseOptions decoder@Dhall.Decoder{expected} mkDef = do
             \ standard output; if '--platform=all' is specified then all pages\
             \ in all platforms are listed"
         )
+      where
+        listAction lang platform sources cfg =
+            Execute cfg (List platform lang sources)
 
     clearCacheFlag :: Options.Parser
         (  Maybe Locale
@@ -305,6 +359,9 @@ parseOptions decoder@Dhall.Decoder{expected} mkDef = do
             \ '--language=LANGUAGE' are specified then they limit what parts\
             \ of the cache are removed."
         )
+      where
+        clearCacheAction lang platform sources cfg =
+            Execute cfg (ClearCache platform lang sources)
 
     platformOption :: Options.Parser SomePlatform
     platformOption = Options.option (Options.eitherReader parsePlatform)
@@ -343,6 +400,8 @@ parseOptions decoder@Dhall.Decoder{expected} mkDef = do
             \ '--sources=SOURCE' is specified only cache for those page\
             \ SOURCEs is updated"
         )
+      where
+        updateAction sources cfg = Execute cfg (Update sources)
 
     sourceOption :: Options.Parser Text
     sourceOption = Options.strOption
@@ -353,3 +412,258 @@ parseOptions decoder@Dhall.Decoder{expected} mkDef = do
             \ SOURCEs; by default all sources are used; this option can be\
             \ used multiple times to specify multiple SOURCEs"
         )
+
+    completionInfoFlag
+        :: Options.Parser Mode
+    completionInfoFlag = Options.flag' CompletionInfo
+        ( Options.long "completion-info"
+        <> Options.help "Describe how command-line completion works in the\
+            \ form of a  Dhall experssion"
+        <> Options.internal
+        )
+
+    completionFlag
+        :: Options.Parser (Shell -> Maybe Word -> [Text] -> Maybe Text -> Mode)
+    completionFlag = Options.flag' completionMode
+        ( Options.long "completion"
+        <> Options.help "Provide command-line completion"
+        <> Options.internal
+        )
+      where
+        completionMode shell index words config =
+            Completion config shell index words
+
+    indexOption :: Options.Parser Word
+    indexOption = Options.option Options.auto
+        ( Options.long "index"
+        <> Options.help "Index of a WORD that we want completion for"
+        <> Options.metavar "INDEX"
+        <> Options.internal
+        )
+
+    shellOption :: Options.Parser Shell
+    shellOption = Options.option parseShell
+        ( Options.long "shell"
+        <> Options.help "Shell for which we want completion for"
+        <> Options.metavar "SHELL"
+        <> Options.internal
+        )
+      where
+        parseShell :: Options.ReadM Shell
+        parseShell = Options.str @(CI String) >>= \case
+            "bash" ->
+                pure Bash
+            "fish" ->
+                pure Fish
+            "zsh" ->
+                pure Zsh
+            _ ->
+                fail "Unrecognised shell name, expected 'bash', 'fish', or\
+                    \ 'zsh'"
+
+    wordOption :: Options.Parser Text
+    wordOption = Options.strOption
+        ( Options.long "word"
+        <> Options.metavar "WORD"
+        <> Options.internal
+        )
+
+data Shell = Bash | Fish | Zsh
+
+completer :: Configuration -> Shell -> Maybe Word -> [Text] -> IO ()
+completer config@Configuration{sources} _shell index words
+  | previousOneOf ["--platform", "-p"] =
+        completePlatform "" current
+
+  | previousOneOf ["--language", "-L"] =
+        completeLanguage "" current
+
+  | previousOneOf ["--source", "-s"] =
+        completeSource "" current
+
+  | previous == Just "--config" =
+        completeConfig "" current
+
+  | Just ('-', _) <- Text.uncons current = if
+      | "--platform=" `Text.isPrefixOf` current ->
+            completePlatform "--platform=" current
+
+      | "--language=" `Text.isPrefixOf` current ->
+            completeLanguage "--language=" current
+
+      | "--source=" `Text.isPrefixOf` current ->
+            completeSource "--source=" current
+
+      | "--config=" `Text.isPrefixOf` current ->
+            completeConfig "--config=" current
+
+        -- Value of `current` is the first option on the command-line.
+      | null before -> do
+            traverse_ Text.putStrLn (prefixMatch current topLevelOptions)
+
+        -- These options mean that nothing else should be completed.
+      | hadBeforeOneOf topLevelTerminalOptions ->
+            pure ()
+
+      | hadBefore "--config-typecheck" -> do
+            let possibilities :: [Text]
+                possibilities = do
+                    let opt = "--config="
+                    opt <$ guard (not (hadBefore opt))
+            traverse_ Text.putStrLn (prefixMatch current possibilities)
+
+      | hadBeforeOneOf ["--update", "-u"] -> do
+            let possibilities :: [Text]
+                possibilities = List.concat
+                    [ ["--source=", "-s"]
+                    , do
+                        guard (not (hadBeforePrefix "--config="))
+                        guard (not (hadBefore "--config"))
+                        pure "--config="
+                    ]
+            traverse_ Text.putStrLn (prefixMatch current possibilities)
+
+        -- "--list", "-l", "--clear-cache", or default mode:
+      | otherwise -> do
+            let possibilities :: [Text]
+                possibilities = List.concat
+                    [ ["--source=", "-s"]
+                    , do
+                        guard (not (hadBeforePrefix "--config="))
+                        guard (not (hadBefore "--config"))
+                        pure "--config="
+                    , do
+                        guard (not (hadBeforePrefix "--language=")
+                        guard (not (hadBeforeOneOf ["--language", "-L"]))
+                        ["--language=", "-L"]
+                    , do
+                        guard (not (hadBeforePrefix "--platform="))
+                        guard (not (hadBeforeOneOf ["--platform", "-p"]))
+                        ["--platform=", "-p"]
+                    ]
+            traverse_ Text.putStrLn (prefixMatch current possibilities)
+
+  | hadBeforeOneOf notDefaultModeOptions =
+      -- There are not arguments, only options in these modes.
+      pure ()
+
+  | otherwise =
+        completeArgument
+            ( List.filter (not . ("-" `Text.isPrefixOf`)) before
+            <> [current]
+            )
+  where
+    before :: [Text]
+    before
+      | null words = []
+      | otherwise  = maybe [] (\i -> List.take (fromIntegral i) words) index
+
+    previous :: Maybe Text
+    previous = lastMay before
+
+    current :: Text
+    current = maybe (lastDef "" words) (atDef "" words . fromIntegral) index
+
+    previousOneOf :: [Text] -> Bool
+    previousOneOf opts = maybe False (`List.elem` opts) previous
+
+    hadBefore :: Text -> Bool
+    hadBefore opt = opt `List.elem` before
+
+    hadBeforeOneOf :: [Text] -> Bool
+    hadBeforeOneOf opts = any (`List.elem` opts) before
+
+    hadBeforePrefix :: Text -> Bool
+    hadBeforePrefix prefix = any (prefix `Text.isPrefixOf`) before
+
+    topLevelOptions :: [Text]
+    topLevelOptions =
+        [ "--help", "-h"
+        , "--version", "-v"
+        , "--config-print-type"
+        , "--config-typecheck"
+        , "--update", "-u"
+        , "--list", "-l"
+        , "--clear-cache"
+        , "--language=", "-L"
+        , "--platform=", "-p"
+        , "--source=", "-s"
+        , "--config="
+        ]
+
+    topLevelTerminalOptions :: [Text]
+    topLevelTerminalOptions =
+        [ "--help", "-h"
+        , "--version", "-v"
+        , "--config-print-type"
+        ]
+
+    notDefaultModeOptions :: [Text]
+    notDefaultModeOptions =
+        [ "--help", "-h"
+        , "--version", "-v"
+        , "--config-print-type"
+        , "--config-typecheck"
+        , "--update", "-u"
+        , "--list", "-l"
+        , "--clear-cache"
+        ]
+
+    prefixMatch :: Text -> [Text] -> [Text]
+    prefixMatch prefix options =
+        List.sort (List.filter (prefix `Text.isPrefixOf`) options)
+
+    completeLanguage :: Text -> Text -> IO ()
+    completeLanguage = withPrefix \word -> do
+        cacheDir <- getCacheDirectory config
+        Index.getIndexFile cacheDir >>= \case
+            Nothing ->
+                pure []
+            Just indexFile ->
+                SQLite.withConnection indexFile \connection ->
+                    List.sort <$> Index.getLocales connection word
+
+    completePlatform :: Text -> Text -> IO ()
+    completePlatform = withPrefix \word -> do
+        let extra = prefixMatch word ["all"]
+        cacheDir <- getCacheDirectory config
+        list <- Index.getIndexFile cacheDir >>= \case
+            Nothing ->
+                pure []
+            Just indexFile ->
+                SQLite.withConnection indexFile \connection ->
+                    Index.getPlatforms connection word
+        pure (List.sort (extra <> list))
+
+    completeSource :: Text -> Text -> IO ()
+    completeSource = withPrefix \word ->
+        pure . prefixMatch word $ NonEmpty.toList sources <&> \Source{name} ->
+            name
+
+    -- TODO: We should support file completion when it starts with one of the
+    -- following characters: '~', '.', '/'. We also have to preserve '~'
+    -- character.
+    completeConfig :: Text -> Text -> IO ()
+    completeConfig = withPrefix \_ -> pure []
+
+    completeArgument :: [Text] -> IO ()
+    completeArgument cmds = do
+        cacheDir <- getCacheDirectory config
+        Index.getIndexFile cacheDir >>= \case
+            Nothing ->
+                pure ()
+            Just indexFile -> do
+                list <- SQLite.withConnection indexFile \connection -> do
+                    let cmd = Text.intercalate "-" cmds
+                    List.sort <$> Index.getCommands connection cmd
+                for_ list \completion -> do
+                    let prefix = Text.intercalate "-"
+                            (maybe [] (<> [""]) (initMay cmds))
+                    Text.putStrLn (Text.drop (Text.length prefix) completion)
+
+    withPrefix :: (Text -> IO [Text]) -> Text -> Text -> IO ()
+    withPrefix f prefix word = do
+        completions <- f (Text.drop (Text.length prefix) word)
+        for_ completions \completion -> do
+            Text.putStr prefix
+            Text.putStrLn completion